+++ src/utils/string/index.ts
@@ -1,3 +1,4 @@
 export { capitalize } from './capitalize/capitalize';
+export { findGroupsWithChar } from './findGroupsWithChar/findGroupsWithChar';
 export { hashString } from './hashString/hashString';
 export { skipBalanced } from './skipBalanced/skipBalanced';